finish with the first round of styling
