+++ src/components/App.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 import MainContainer from '../components/MainContainer';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import NavContainer from './NavContainer';
-import FeaturesPageContainer from './FeaturesPageContainer'
+import AboutPageContainer from './AboutPageContainer'
 
 export default function App(){
 
   return (
     <div>
     <Router>
-    <NavContainer/>
+      <NavContainer/>
       <Routes>
           <Route path="/" element={<MainContainer/>}></Route>
-          <Route path="/about" element={<FeaturesPageContainer/>}></Route>
+          <Route path="/about" element={<AboutPageContainer/>}></Route>
       </Routes>
     </Router>
     </div>
