+++ src/components/NavContainer.jsx
@@ -18,14 +18,14 @@ export default function NavContainer(){
 
     return (
         <div className='navContainer'>
-          <div id='logo_container'>
           <img id='image-logo-main' src={logo} />
+          <div id="links">
+            <Link to='/' className='nav-link'>Home</Link>
+            <Link to='/about' className='nav-link'>About</Link>
+            <Link to='https://github.com/oslabs-beta/K8sDependencyTracker' target="_blank" className='nav-link'>GitHub</Link>
+            <Link to='/' className='nav-link' onClick={() => setShouldScrollToBottom(true)}>Team</Link>
+            <Link to='https://medium.com/@mhart1992/how-to-track-k8s-api-versioning-aac19f85e94d' target="_blank" className='nav-link'>Medium</Link>
           </div>
-          <Link to='/' className='nav-link'>Home</Link>
-          <Link to='/about' className='nav-link'>About</Link>
-          <Link to='https://github.com/oslabs-beta/K8sDependencyTracker' target="_blank" className='nav-link'>GitHub</Link>
-          <Link to='/' className='nav-link' onClick={() => setShouldScrollToBottom(true)}>Team</Link>
-          <Link to='https://medium.com/@mhart1992/how-to-track-k8s-api-versioning-aac19f85e94d' target="_blank" className='nav-link'>Medium</Link>
         </div>
       )
   };
